Prepare to release v4.1.0.
This release moves the CLI binary out to an example so we can lose a lot of dependencies that aren't needed to use credential stores.

In addition, it recreates the v1 (and v2/v3) functionality of providing an `Entry` object that any app can use (without special features or initialization) to manage text and binary secrets. This is provided by the `v1` feature, which is the default. (This change fixes #317.)

The ability to use any credential store is still preserved in the `cli` module but the `cli` feature which exposes it is not in the default feature set.

Strictly speaking, these changes are probably SemVer-significant. But since there are few to no clients of the existing library functionality, we're calling this a patch release.
