+++ .github/workflows/ci.yaml
@@ -2,11 +2,7 @@ name: CI
 
 on:
   push:
-    paths-ignore:
-      - ./python/**
   pull_request:
-    paths-ignore:
-      - ./python/**
   workflow_dispatch:
 
 jobs:
