+++ examples/cli/main.rs
